move queryParams to searcher
