+++ src/common/Searcher/searchQueryParamName.js
@@ -1,0 +1,4 @@
+
+export const searchQueryParamName = "query";
+export const searchTypePage = "type";
+export const searchPageNumber = "page";
