+++ src/common/Searcher/useQueryParameter.js
@@ -1,8 +1,8 @@
 import { useLocation } from "react-router-dom";
 
 const useQueryParameter = (paramName) => {
-const location = useLocation();
-return (new URLSearchParams(location.search)).get(paramName);
+    const location = useLocation();
+    return (new URLSearchParams(location.search)).get(paramName);
 }
 
 export default useQueryParameter;
