--- src/common/searchQueryParamName.js
@@ -1,4 +1,0 @@
-
-const searchQueryParamName = "szukaj";
-  
-export default searchQueryParamName;
